Add magnifier button to the MainWindow.axaml
Description:

This commit adds a magnifier button to the MainWindow.axaml.
The button's purpose is to enable future search functionality
for workflows or specific words within a workflow or an
'Erweiterte Eigenschaft' (expanded property).

Context:

During an update/switch of a shop in JTL WaWi, I encountered
the need to lookup all workflows that might break when
switching names or making other changes to the shop.