Add searchable texts to the workflow
Description:

Workflows can have conditions and actions, both of which are
represented as strings/XML. To enable searching for values in
the workflows, we need to add them to the Workflow objects.

Context:

This is a straightforward way to make searching possible, but
it has some drawbacks. In addition to the strings we want to
search through, we also include all the XML overhead, which
can lead to false positives. Since we will encapsulate the
searching functionality, we can later switch from this simple
string crawling to a more sophisticated solution.